разбивает js на модули и подключает к странице
